# Optimize Crc32 for arm