--- a/tests/ImageSharp.Tests/TestUtilities/Tests/FeatureTestRunnerTests.cs
+++ b/tests/ImageSharp.Tests/TestUtilities/Tests/FeatureTestRunnerTests.cs
@@ -7,6 +7,10 @@ using System.Linq;
 using System.Numerics;
 #if SUPPORTS_RUNTIME_INTRINSICS
 using System.Runtime.Intrinsics.X86;
+using Aes = System.Runtime.Intrinsics.X86.Aes;
+#if NET5_0_OR_GREATER
+using System.Runtime.Intrinsics.Arm;
+#endif
 #endif
 using Xunit;
 using Xunit.Abstractions;
@@ -16,11 +20,11 @@ namespace SixLabors.ImageSharp.Tests.TestUtilities.Tests
     public class FeatureTestRunnerTests
     {
         public static TheoryData<HwIntrinsics, string[]> Intrinsics =>
-            new TheoryData<HwIntrinsics, string[]>
+            new()
             {
-                { HwIntrinsics.DisableAES | HwIntrinsics.AllowAll, new string[] { "EnableAES", "AllowAll" } },
-                { HwIntrinsics.DisableSIMD | HwIntrinsics.DisableHWIntrinsic, new string[] { "FeatureSIMD", "EnableHWIntrinsic" } },
-                { HwIntrinsics.DisableSSE42 | HwIntrinsics.DisableAVX, new string[] { "EnableSSE42", "EnableAVX" } }
+                { HwIntrinsics.DisableAES | HwIntrinsics.AllowAll, new[] { "EnableAES", "AllowAll" } },
+                { HwIntrinsics.DisableSIMD | HwIntrinsics.DisableHWIntrinsic, new[] { "FeatureSIMD", "EnableHWIntrinsic" } },
+                { HwIntrinsics.DisableSSE42 | HwIntrinsics.DisableAVX, new[] { "EnableSSE42", "EnableAVX" } }
             };
 
         [Theory]
@@ -56,12 +60,9 @@ namespace SixLabors.ImageSharp.Tests.TestUtilities.Tests
         }
 
         [Fact]
-        public void CanLimitHwIntrinsicSIMDFeatures()
-        {
-            FeatureTestRunner.RunWithHwIntrinsicsFeature(
+        public void CanLimitHwIntrinsicSIMDFeatures() => FeatureTestRunner.RunWithHwIntrinsicsFeature(
                 () => Assert.False(Vector.IsHardwareAccelerated),
                 HwIntrinsics.DisableSIMD);
-        }
 
 #if SUPPORTS_RUNTIME_INTRINSICS
         [Fact]
@@ -121,6 +122,14 @@ namespace SixLabors.ImageSharp.Tests.TestUtilities.Tests
                         Assert.False(Bmi1.IsSupported);
                         Assert.False(Bmi2.IsSupported);
                         Assert.False(Lzcnt.IsSupported);
+#if NET5_0_OR_GREATER
+                        Assert.False(AdvSimd.IsSupported);
+                        Assert.False(System.Runtime.Intrinsics.Arm.Aes.IsSupported);
+                        Assert.False(Crc32.IsSupported);
+                        Assert.False(Dp.IsSupported);
+                        Assert.False(Sha1.IsSupported);
+                        Assert.False(Sha256.IsSupported);
+#endif
                         break;
                     case HwIntrinsics.DisableSSE:
                         Assert.False(Sse.IsSupported);
@@ -167,6 +176,26 @@ namespace SixLabors.ImageSharp.Tests.TestUtilities.Tests
                     case HwIntrinsics.DisableLZCNT:
                         Assert.False(Lzcnt.IsSupported);
                         break;
+#if NET5_0_OR_GREATER
+                    case HwIntrinsics.DisableArm64AdvSimd:
+                        Assert.False(AdvSimd.IsSupported);
+                        break;
+                    case HwIntrinsics.DisableArm64Aes:
+                        Assert.False(System.Runtime.Intrinsics.Arm.Aes.IsSupported);
+                        break;
+                    case HwIntrinsics.DisableArm64Crc32:
+                        Assert.False(Crc32.IsSupported);
+                        break;
+                    case HwIntrinsics.DisableArm64Dp:
+                        Assert.False(Dp.IsSupported);
+                        break;
+                    case HwIntrinsics.DisableArm64Sha1:
+                        Assert.False(Sha1.IsSupported);
+                        break;
+                    case HwIntrinsics.DisableArm64Sha256:
+                        Assert.False(Sha256.IsSupported);
+                        break;
+#endif
 #endif
                 }
             }
@@ -226,6 +255,14 @@ namespace SixLabors.ImageSharp.Tests.TestUtilities.Tests
                         Assert.False(Bmi1.IsSupported);
                         Assert.False(Bmi2.IsSupported);
                         Assert.False(Lzcnt.IsSupported);
+#if NET5_0_OR_GREATER
+                        Assert.False(AdvSimd.IsSupported);
+                        Assert.False(System.Runtime.Intrinsics.Arm.Aes.IsSupported);
+                        Assert.False(Crc32.IsSupported);
+                        Assert.False(Dp.IsSupported);
+                        Assert.False(Sha1.IsSupported);
+                        Assert.False(Sha256.IsSupported);
+#endif
                         break;
                     case HwIntrinsics.DisableSSE:
                         Assert.False(Sse.IsSupported);
@@ -272,6 +309,26 @@ namespace SixLabors.ImageSharp.Tests.TestUtilities.Tests
                     case HwIntrinsics.DisableLZCNT:
                         Assert.False(Lzcnt.IsSupported);
                         break;
+#if NET5_0_OR_GREATER
+                    case HwIntrinsics.DisableArm64AdvSimd:
+                        Assert.False(AdvSimd.IsSupported);
+                        break;
+                    case HwIntrinsics.DisableArm64Aes:
+                        Assert.False(System.Runtime.Intrinsics.Arm.Aes.IsSupported);
+                        break;
+                    case HwIntrinsics.DisableArm64Crc32:
+                        Assert.False(Crc32.IsSupported);
+                        break;
+                    case HwIntrinsics.DisableArm64Dp:
+                        Assert.False(Dp.IsSupported);
+                        break;
+                    case HwIntrinsics.DisableArm64Sha1:
+                        Assert.False(Sha1.IsSupported);
+                        break;
+                    case HwIntrinsics.DisableArm64Sha256:
+                        Assert.False(Sha256.IsSupported);
+                        break;
+#endif
 #endif
                 }
             }