### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
<!-- A description of the changes proposed in the pull-request -->
Adds a fast path in Crc32 for arm and arm64

Some benchmarks for the arm64 path:

Current (LUT):
|             Method | Count |     Mean |    Error |   StdDev |
|------------------- |------ |---------:|---------:|---------:|
| SixLaborsCalculate |  1024 | 14.54 us | 0.001 us | 0.001 us |
| SixLaborsCalculate |  2048 | 29.07 us | 0.003 us | 0.002 us |
| SixLaborsCalculate |  4096 | 58.14 us | 0.008 us | 0.008 us |

arm crc32 instructions without loop unrolling:
|             Method | Count |       Mean |   Error |  StdDev |
|------------------- |------ |-----------:|--------:|--------:|
| SixLaborsCalculate |  1024 |   833.0 ns | 7.47 ns | 6.62 ns |
| SixLaborsCalculate |  2048 | 1,557.8 ns | 0.17 ns | 0.14 ns |
| SixLaborsCalculate |  4096 | 3,044.1 ns | 0.75 ns | 0.70 ns |

This PR:
|             Method | Count |       Mean |   Error |  StdDev |
|------------------- |------ |-----------:|--------:|--------:|
| SixLaborsCalculate |  1024 |   584.8 ns | 0.07 ns | 0.06 ns |
| SixLaborsCalculate |  2048 | 1,145.8 ns | 0.12 ns | 0.11 ns |
| SixLaborsCalculate |  4096 | 2,262.1 ns | 0.35 ns | 0.29 ns |
<!-- Thanks for contributing to ImageSharp! -->