@antonfirsov 2022-02-11
Hmmm .. this is difficult.

On one hand I really appreciate and value the contribution, and would be happy to see us getting better on ARM. On the other hand I'm afraid we lack some prerequisites that are necessary to take it in:
1. We don't run CI tests against ARM, and we don't have easy access to ARM in other ways. This means that there is no way to maintain this optimization guaranteeing correctness.
2. The .NET 5/6 targets are test-only meaning they do not actually make their ways to our NuGet package, the highest version is still 3.1. We will update as part of our 3.0 roadmap, but it's not high priority right now, so I don't think it will happen very soon.

As such I'm leaning towards not merging this to `main` but shelving it to a separate branch for future adaption.

@SixLabors/core thoughts?