@JimBobSquarePants 2022-02-11
@antonfirsov Can we use this to emulate the environment with QEMU?
https://github.com/uraimo/run-on-arch-action