@brianpopow 2022-02-11
> @antonfirsov Can we use this to emulate the environment with QEMU?
https://github.com/uraimo/run-on-arch-action

For this to work we would need also a docker image with ImageMagick build for ARM, otherwise most of the test wont work.
That's why I was experimenting building ImageMagick for ARM a while ago.

While this can work, it will be very slow. My very rough estimate would be, that running all test with QEMU would take something like 40 minutes (just a wild guess, but i doubt it will be much faster).

Also what I am not sure about is, if QEMU will emulate hardware intrinsics, too.

In my opinion we should keep the changes from this PR in a separate branch until we figure out how to run the tests properly on the CI.