@JimBobSquarePants 2022-02-11
@dlemstra got ARM builds working I think. By my understanding QEMU does emulate hardware intrinsics but I don’t know if there’s limitations.

Alternatively we can look at purchasing a device and wiring up CI to talk to that (I’ve read that’s possible but no idea how). 

Re branching… yeah. Very wise.