@brianpopow 2022-02-11
> Alternatively we can look at purchasing a device and wiring up CI to talk to that (I’ve read that’s possible but no idea how).

Yes a selfhosted CI runner would work, but this device then must run all the time somewhere and always be reachable to the internet. Not sure how feasible that would be.