@JimBobSquarePants 2022-02-11
I have a static IP so possible… 

I’d like to explore QEMU first though