@brianpopow 2022-02-12
I have tried out https://github.com/uraimo/run-on-arch-action with a simple example.
So far without success. When I run any `dotnet` command, I get:

```
Fatal error. System.Runtime.InteropServices.SEHException (0x80004005): External component has thrown an exception.
   at System.DateTime.op_Subtraction(System.DateTime, System.DateTime)
   at Microsoft.DotNet.Cli.Program.Main(System.String[])
```

Besides that, I have tried running a simple unit test which uses ARM intrinsics with docker (which also uses QEMU when the host is not ARM) and it seems to work. So in theory QEMU should work. Not sure why `run-on-arch-action` does not work. I guess it uses a different version of QEMU.