@Bond-009 2022-02-14
@JimBobSquarePants We don't have automated arm tests for jellyfin. (We also don't have any arm specific code)

Automated tests would only help with correctness regressions, not performance regressions, so you'd still have to check that manually.