@JimBobSquarePants 2022-02-14
>Automated tests would only help with correctness regressions, not performance regressions, so you'd still have to check that manually.

We've actually developed a technique here that allows us to use automated testing against intrinsics features. We're able to switch off intrinsics features to ensure correctness for all implementations and fallbacks.

https://github.com/SixLabors/ImageSharp/blob/46b0bf0b495ec435fc404e014854d8ca461e7f7a/tests/ImageSharp.Tests/TestUtilities/FeatureTesting/FeatureTestRunner.cs#L52