@brianpopow 2022-02-14
Maybe we should adjust the `Crc32Tests` to use the `FeatureTestRunner`, its currently not.