@JimBobSquarePants 2022-02-14
Good idea. I need to check the runtime also to see if there’s additional flags now for arm intrinsics