@JimBobSquarePants 2022-02-16
Found the new collection of [constants](https://github.com/dotnet/runtime/blob/50ac454d8d8a1915188b2a4bb3fff3b81bf6c0cf/src/coreclr/src/jit/jitconfigvalues.h#L255-L278). We'll have to figure out which ones are most relevant.