@JimBobSquarePants 2022-02-20
@tannergooding Would it be possible for you to advise which [constants](https://github.com/dotnet/runtime/blob/50ac454d8d8a1915188b2a4bb3fff3b81bf6c0cf/src/coreclr/src/jit/jitconfigvalues.h#L255-L278) we should add to our `HwIntrinsics` enum for feature testing ARM intrinsics? Thanks!

https://github.com/SixLabors/ImageSharp/blob/92da254bb8203ea5eb2fc8f58f849f487921466e/tests/ImageSharp.Tests/TestUtilities/FeatureTesting/FeatureTestRunner.cs#L388-L410